fix: cast every sys key constant to i32 in the enum discriminants to be cross‑platform

+++ dear-imgui/src/input.rs
@@ -44,151 +44,151 @@ pub enum MouseCursor {
 #[derive(Copy, Clone, Debug, Hash, Eq, PartialEq)]
 pub enum Key {
     /// No key
-    None = sys::ImGuiKey_None,
+    None = sys::ImGuiKey_None as i32,
     /// Tab key
-    Tab = sys::ImGuiKey_Tab,
+    Tab = sys::ImGuiKey_Tab as i32,
     /// Left arrow key
-    LeftArrow = sys::ImGuiKey_LeftArrow,
+    LeftArrow = sys::ImGuiKey_LeftArrow as i32,
     /// Right arrow key
-    RightArrow = sys::ImGuiKey_RightArrow,
+    RightArrow = sys::ImGuiKey_RightArrow as i32,
     /// Up arrow key
-    UpArrow = sys::ImGuiKey_UpArrow,
+    UpArrow = sys::ImGuiKey_UpArrow as i32,
     /// Down arrow key
-    DownArrow = sys::ImGuiKey_DownArrow,
+    DownArrow = sys::ImGuiKey_DownArrow as i32,
     /// Page up key
-    PageUp = sys::ImGuiKey_PageUp,
+    PageUp = sys::ImGuiKey_PageUp as i32,
     /// Page down key
-    PageDown = sys::ImGuiKey_PageDown,
+    PageDown = sys::ImGuiKey_PageDown as i32,
     /// Home key
-    Home = sys::ImGuiKey_Home,
+    Home = sys::ImGuiKey_Home as i32,
     /// End key
-    End = sys::ImGuiKey_End,
+    End = sys::ImGuiKey_End as i32,
     /// Insert key
-    Insert = sys::ImGuiKey_Insert,
+    Insert = sys::ImGuiKey_Insert as i32,
     /// Delete key
-    Delete = sys::ImGuiKey_Delete,
+    Delete = sys::ImGuiKey_Delete as i32,
     /// Backspace key
-    Backspace = sys::ImGuiKey_Backspace,
+    Backspace = sys::ImGuiKey_Backspace as i32,
     /// Space key
-    Space = sys::ImGuiKey_Space,
+    Space = sys::ImGuiKey_Space as i32,
     /// Enter key
-    Enter = sys::ImGuiKey_Enter,
+    Enter = sys::ImGuiKey_Enter as i32,
     /// Escape key
-    Escape = sys::ImGuiKey_Escape,
+    Escape = sys::ImGuiKey_Escape as i32,
     /// Left Ctrl key
-    LeftCtrl = sys::ImGuiKey_LeftCtrl,
+    LeftCtrl = sys::ImGuiKey_LeftCtrl as i32,
     /// Left Shift key
-    LeftShift = sys::ImGuiKey_LeftShift,
+    LeftShift = sys::ImGuiKey_LeftShift as i32,
     /// Left Alt key
-    LeftAlt = sys::ImGuiKey_LeftAlt,
+    LeftAlt = sys::ImGuiKey_LeftAlt as i32,
     /// Left Super key
-    LeftSuper = sys::ImGuiKey_LeftSuper,
+    LeftSuper = sys::ImGuiKey_LeftSuper as i32,
     /// Right Ctrl key
-    RightCtrl = sys::ImGuiKey_RightCtrl,
+    RightCtrl = sys::ImGuiKey_RightCtrl as i32,
     /// Right Shift key
-    RightShift = sys::ImGuiKey_RightShift,
+    RightShift = sys::ImGuiKey_RightShift as i32,
     /// Right Alt key
-    RightAlt = sys::ImGuiKey_RightAlt,
+    RightAlt = sys::ImGuiKey_RightAlt as i32,
     /// Right Super key
-    RightSuper = sys::ImGuiKey_RightSuper,
+    RightSuper = sys::ImGuiKey_RightSuper as i32,
     /// Menu key
-    Menu = sys::ImGuiKey_Menu,
+    Menu = sys::ImGuiKey_Menu as i32,
     /// 0 key
-    Key0 = sys::ImGuiKey_0,
+    Key0 = sys::ImGuiKey_0 as i32,
     /// 1 key
-    Key1 = sys::ImGuiKey_1,
+    Key1 = sys::ImGuiKey_1 as i32,
     /// 2 key
-    Key2 = sys::ImGuiKey_2,
+    Key2 = sys::ImGuiKey_2 as i32,
     /// 3 key
-    Key3 = sys::ImGuiKey_3,
+    Key3 = sys::ImGuiKey_3 as i32,
     /// 4 key
-    Key4 = sys::ImGuiKey_4,
+    Key4 = sys::ImGuiKey_4 as i32,
     /// 5 key
-    Key5 = sys::ImGuiKey_5,
+    Key5 = sys::ImGuiKey_5 as i32,
     /// 6 key
-    Key6 = sys::ImGuiKey_6,
+    Key6 = sys::ImGuiKey_6 as i32,
     /// 7 key
-    Key7 = sys::ImGuiKey_7,
+    Key7 = sys::ImGuiKey_7 as i32,
     /// 8 key
-    Key8 = sys::ImGuiKey_8,
+    Key8 = sys::ImGuiKey_8 as i32,
     /// 9 key
-    Key9 = sys::ImGuiKey_9,
+    Key9 = sys::ImGuiKey_9 as i32,
     /// A key
-    A = sys::ImGuiKey_A,
+    A = sys::ImGuiKey_A as i32,
     /// B key
-    B = sys::ImGuiKey_B,
+    B = sys::ImGuiKey_B as i32,
     /// C key
-    C = sys::ImGuiKey_C,
+    C = sys::ImGuiKey_C as i32,
     /// D key
-    D = sys::ImGuiKey_D,
+    D = sys::ImGuiKey_D as i32,
     /// E key
-    E = sys::ImGuiKey_E,
+    E = sys::ImGuiKey_E as i32,
     /// F key
-    F = sys::ImGuiKey_F,
+    F = sys::ImGuiKey_F as i32,
     /// G key
-    G = sys::ImGuiKey_G,
+    G = sys::ImGuiKey_G as i32,
     /// H key
-    H = sys::ImGuiKey_H,
+    H = sys::ImGuiKey_H as i32,
     /// I key
-    I = sys::ImGuiKey_I,
+    I = sys::ImGuiKey_I as i32,
     /// J key
-    J = sys::ImGuiKey_J,
+    J = sys::ImGuiKey_J as i32,
     /// K key
-    K = sys::ImGuiKey_K,
+    K = sys::ImGuiKey_K as i32,
     /// L key
-    L = sys::ImGuiKey_L,
+    L = sys::ImGuiKey_L as i32,
     /// M key
-    M = sys::ImGuiKey_M,
+    M = sys::ImGuiKey_M as i32,
     /// N key
-    N = sys::ImGuiKey_N,
+    N = sys::ImGuiKey_N as i32,
     /// O key
-    O = sys::ImGuiKey_O,
+    O = sys::ImGuiKey_O as i32,
     /// P key
-    P = sys::ImGuiKey_P,
+    P = sys::ImGuiKey_P as i32,
     /// Q key
-    Q = sys::ImGuiKey_Q,
+    Q = sys::ImGuiKey_Q as i32,
     /// R key
-    R = sys::ImGuiKey_R,
+    R = sys::ImGuiKey_R as i32,
     /// S key
-    S = sys::ImGuiKey_S,
+    S = sys::ImGuiKey_S as i32,
     /// T key
-    T = sys::ImGuiKey_T,
+    T = sys::ImGuiKey_T as i32,
     /// U key
-    U = sys::ImGuiKey_U,
+    U = sys::ImGuiKey_U as i32,
     /// V key
-    V = sys::ImGuiKey_V,
+    V = sys::ImGuiKey_V as i32,
     /// W key
-    W = sys::ImGuiKey_W,
+    W = sys::ImGuiKey_W as i32,
     /// X key
-    X = sys::ImGuiKey_X,
+    X = sys::ImGuiKey_X as i32,
     /// Y key
-    Y = sys::ImGuiKey_Y,
+    Y = sys::ImGuiKey_Y as i32,
     /// Z key
-    Z = sys::ImGuiKey_Z,
+    Z = sys::ImGuiKey_Z as i32,
     /// F1 key
-    F1 = sys::ImGuiKey_F1,
+    F1 = sys::ImGuiKey_F1 as i32,
     /// F2 key
-    F2 = sys::ImGuiKey_F2,
+    F2 = sys::ImGuiKey_F2 as i32,
     /// F3 key
-    F3 = sys::ImGuiKey_F3,
+    F3 = sys::ImGuiKey_F3 as i32,
     /// F4 key
-    F4 = sys::ImGuiKey_F4,
+    F4 = sys::ImGuiKey_F4 as i32,
     /// F5 key
-    F5 = sys::ImGuiKey_F5,
+    F5 = sys::ImGuiKey_F5 as i32,
     /// F6 key
-    F6 = sys::ImGuiKey_F6,
+    F6 = sys::ImGuiKey_F6 as i32,
     /// F7 key
-    F7 = sys::ImGuiKey_F7,
+    F7 = sys::ImGuiKey_F7 as i32,
     /// F8 key
-    F8 = sys::ImGuiKey_F8,
+    F8 = sys::ImGuiKey_F8 as i32,
     /// F9 key
-    F9 = sys::ImGuiKey_F9,
+    F9 = sys::ImGuiKey_F9 as i32,
     /// F10 key
-    F10 = sys::ImGuiKey_F10,
+    F10 = sys::ImGuiKey_F10 as i32,
     /// F11 key
-    F11 = sys::ImGuiKey_F11,
+    F11 = sys::ImGuiKey_F11 as i32,
     /// F12 key
-    F12 = sys::ImGuiKey_F12,
+    F12 = sys::ImGuiKey_F12 as i32,
 }
 
 impl From<MouseButton> for sys::ImGuiMouseButton {
